requirements.txt
Flask==1.1.1
redis==3.2.1
opencv_python==4.0.1.24
numpy==1.17.2
Pillow==7.0.0